+++ model/order.go
@@ -3,7 +3,7 @@ package model
 // Order base model.
 type Order struct {
 	Number     string
-	UserID     int
+	UserID     uint64
 	Status     string
 	Accrual    float64
 	UploadedAT string
