+++ service/interface.go
@@ -10,14 +10,14 @@ import (
 type UserProcessor interface {
 	Create(ctx context.Context, login, password string) error
 	Auth(ctx context.Context, login, password string) (string, error)
-	ShowBalance(ctx context.Context, userID int) (*response.Balance, error)
-	GetTransactions(ctx context.Context, userID int) ([]model.Transaction, error)
-	Withdraw(ctx context.Context, userID int, orderID string, sum float64) error
+	ShowBalance(ctx context.Context, userID uint64) (*response.Balance, error)
+	GetTransactions(ctx context.Context, userID uint64) ([]model.Transaction, error)
+	Withdraw(ctx context.Context, userID uint64, orderID string, sum float64) error
 }
 
 // OrderProcessor interface for operations with orders.
 type OrderProcessor interface {
-	CreateOrder(ctx context.Context, number string, userID int) error
+	CreateOrder(ctx context.Context, number string, userID uint64) error
 	GetOrderByNumber(ctx context.Context, number string) (*model.Order, error)
-	GetOrdersByUser(ctx context.Context, userID int) ([]model.Order, error)
+	GetOrdersByUser(ctx context.Context, userID uint64) ([]model.Order, error)
 }
