+++ storage/pgsql/user.go
@@ -37,7 +37,7 @@ func (s Storage) GetUserByLogin(ctx context.Context, login string) (*model.User,
 }
 
 // GetUserByID returns user by login
-func (s Storage) GetUserByID(ctx context.Context, id int) (*model.User, error) {
+func (s Storage) GetUserByID(ctx context.Context, id uint64) (*model.User, error) {
 	var user dbModel.User
 	query := "select * from users where id = $1"
 	if err := s.db.GetContext(ctx, &user, query, id); err != nil {
